backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
requests
python-dotenv
pydantic
pydantic-settings
asyncpg
python-multipart
langsmith
loguru